プログラム全体を調整
